% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSCseq.R
\name{E_step}
\alias{E_step}
\title{Computes E step weights}
\usage{
E_step(wts, l, pi, CEM, Tau, PP_filt)
}
\arguments{
\item{wts}{matrix of dimension k by n, current E step weights. Used as template matrix to update}

\item{l}{matrix of dimension k by n, log-likelihood summed over all genes, for each k clusters and n subjects}

\item{pi}{vector of k mixture proportions}

\item{CEM}{logical, FALSE for EM and TRUE for CEM}

\item{Tau}{numeric, current temperature of CEM. Tau=1 for EM}
}
\value{
list containing updated wts: matrix of E step weights,
keep: logical matrix of whether a sample is included in cluster baseline calculation,
Tau: numeric: resultant CEM temperature (after annealing),
CEM: logical: whether CEM is continued to be run
}
\description{
Takes current parameter estimates and likelihoods, and
computes E step weights. Also updates CEM temperature Tau,
if applicable
}
