% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSCseq_workflow.R
\name{FSCseq_predict_workflow}
\alias{FSCseq_predict_workflow}
\title{Minimal workflow for FSCseq_predict}
\usage{
FSCseq_predict_workflow(X = NULL, fit, cts, cts_pred, idx)
}
\arguments{
\item{X}{covariates (optional)}

\item{fit}{FSCseq results object. Accessed by $results from FSCseq_workflow object}

\item{cts}{Training set count matrix, dimension g by n. Used as pseudo-reference to calculate prediction set size factors}

\item{cts_pred}{integer matrix, count matrix of dimension g by n_pred. Must be integers (counts)}

\item{idx}{boolean vector: TRUE if gene passed pre-filtering step}
}
\value{
list with processed.dat.pred (processed prediction data), and prediction results
}
\description{
Full FSCseq workflow based on minimal working defaults
}
