% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSCseq.R
\name{glm.init}
\alias{glm.init}
\title{Initializes parameter estimates}
\usage{
glm.init(j, y_j, XX, k, offsets, wts, keep)
}
\arguments{
\item{j}{integer, gene index j}

\item{y_j}{vector of length n, expression of gene j}

\item{XX}{matrix of dimension nk by k+p, design matrix}

\item{k}{integer, number of clusters}

\item{offsets}{numeric vector of length n, subject-specific offsets}

\item{wts}{matrix of dimension k by n, E step weights}

\item{keep}{logical matrix of dimension k by n, whether a sample is included in calculation for cluster k}
}
\value{
list containing outputs coefs_j: vector of length k corresponding to each cluster log2 baseline,
phi_g: Overdispersion estimate
}
\description{
Uses glm() and phi_ml_g() functions to initialize
estimates at the beginning of EM algorithm.
}
