% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSCseq.R
\name{glm.init_par}
\alias{glm.init_par}
\title{Parallelized glm.init}
\usage{
glm.init_par(j)
}
\arguments{
\item{j}{integer, gene index j}
}
\value{
list containing outputs coefs_j: vector of length k corresponding to each cluster log2 baseline,
phi_g: Overdispersion estimate
}
\description{
Uses glm() and phi_ml_g() functions to initialize
estimates at the beginning of EM algorithm.
}
