% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSCseq.R
\name{logsumexpc}
\alias{logsumexpc}
\title{Logsumexp technique to prevent underflow}
\usage{
logsumexpc(v)
}
\arguments{
\item{v}{a vector or matrix of numbers}
}
\value{
log(sum(exp(v))), without underflow issues
}
\description{
This function takes in a vector, and uses the logsumexp
technique to prevent underflow issues. Used on
log_pi + loglike to prevent it being rounded to 0
}
\examples{
log_pi=log(c(0.1,0.02,0.07))
l=matrix(c(-2000,-2100,-2050,-1000,-1050,-1100),nrow=3,ncol=2)
logsumexpc(log_pi+l)
log(sum(exp(log_pi+l)))

}
