% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{M_step}
\alias{M_step}
\title{Perform M step of EM algorithm}
\usage{
M_step(X, y_j, p, j, a, k, all_wts, keep, offset, theta, coefs_j, phi_j,
  cl_phi, est_covar, lambda, alpha, IRLS_tol, maxit_IRLS, optim_method)
}
\arguments{
\item{X}{design matrix}
}
\description{
Perform M step of EM algorithm
}
