% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSCseq.R
\name{M_step_par}
\alias{M_step_par}
\title{Parallelized M step}
\usage{
M_step_par(j)
}
\arguments{
\item{j}{gene index}
}
\value{
Same output as M_step() function
}
\description{
Runs M_step() for gene j. Used to compute in parallel on ncores (specified in wrapper)
}
