% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSCseq.R
\name{M_step_par2}
\alias{M_step_par2}
\title{Parallelized M step (mclapply)}
\usage{
M_step_par2(j, XX, y, p, a, k, wts, keep, offsets, theta_list, coefs, phi,
  cl_phi, est_phi, est_covar, lambda, alpha, IRLS_tol, maxit_IRLS, Tau,
  disc_ids_list, par_X)
}
\arguments{
\item{j}{gene index}
}
\value{
Same output as M_step() function
}
\description{
Runs M_step() for gene j. Used to compute in parallel on ncores (specified in wrapper)
}
