% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSCseq.R
\name{phi.ml}
\alias{phi.ml}
\title{Estimates dispersion parameter phi}
\usage{
phi.ml(y, mu, dfr, weights, t0, limit, trace)
}
\arguments{
\item{y}{vector: gene read counts for one fixed gene}

\item{mu}{vector: mean. must be same length as y}

\item{dfr}{numeric: degrees of freedom for theta.mm()}

\item{weights}{vector: must be same length as y}

\item{limit}{numeric: max number of iterations for theta.ml() and theta.mm()}

\item{trace}{logical: True/False to show trace output from theta.ml}
}
\value{
phi_g: estimate of phi
}
\description{
Uses theta.ml2 and theta.mm functions to tryCatch
any unstable estimations
}
