% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{phi_ml_g}
\alias{phi_ml_g}
\title{Perform phi_ml estimation of phi}
\usage{
phi_ml_g(y, mu, wts, limit, p0, trace)
}
\arguments{
\item{y, }{mu, wts, limit, p0, trace}
}
\description{
Perform phi_ml estimation of phi
}
