% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{processData}
\alias{processData}
\title{Pre-process the count matrix}
\usage{
processData(y, geoMeans = NULL, estimateSFtype = "ratio",
  med_filt = TRUE, MAD_filt = TRUE, med_thresh = 100,
  MAD_quant_thresh = 50)
}
\arguments{
\item{y}{integer, gene expression count matrix (output from simulateData)}

\item{geoMeans}{(optional) numeric vector of length g (number of rows of y): custom geometric means of gene counts. Used to estimate prediction set size factors based on training set}

\item{estimateSFtype}{string, input for estimateSizeFactors "type" argument. must be 'ratio', 'poscounts', or 'iterate'. See DESeq2 vignette.}

\item{med_filt}{logical, TRUE to filter low-count genes via median threshold}

\item{MAD_filt}{logical, TRUE to filter low-variable genes via MAD quantile threshold}

\item{med_thresh}{numeric, median threshold for pre-filtering low-count genes (default 100, i.e. pre-filters genes with median normalized count below 100)}

\item{MAD_quant_thresh}{numeric value between 0 to 100, quantile threshold imposed on MAD values for pre-filtering low-variable genes (default 50, i.e. pre-filters genes below 50th quantile of MAD values)}
}
\value{
list containing the following objects:
dds: DESeq2 output.
size_factors: numeric vector, estimated size factors, from DESeq2
norm_y: numeric normalized count matrix, corrected for differences in sequencing depth, from DESeq2
idx: logical vector, TRUE for inclusion after pre-filtering low-count and low-variable genes
row_medians: numeric vector, median normalized count for each gene
row_MADs: numeric vector, median absolute deviation (MAD) value of log(norm_y+0.1) for each gene
}
\description{
Estimate size factors using DESeq2, and pre-filter based on low count and
low variability genes using custom thresholds on median normalized count and
median absolute deviation (MAD) values.
}
