% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateData.R
\name{simulate_counts}
\alias{simulate_counts}
\title{Simulate data with gene-wise dispersion parameters}
\usage{
simulate_counts(K, B, n, g, cls, SF, beta, phi, LFCg_mat, batch, LFCb,
  sigma_b, sigma_g, DEb_ID, disp)
}
\description{
Simulate data with gene-wise dispersion parameters
}
