% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateData.R
\name{simulateData}
\alias{simulateData}
\title{Simulate RNA-seq bulk gene expression count data}
\usage{
simulateData(n, k, g, pi = rep(1/k, k), coefs = NULL, beta, p_disc,
  LFC, disp = "gene", phi, size_factors = NULL, batch_effects = 0,
  batch = rep(1, n))
}
\arguments{
\item{n}{integer, sample size}

\item{k}{integer, number of clusters}

\item{g}{integer, number of genes}

\item{pi}{numeric vector, mixture proportions, or proportion in each cluster. Equal sizes by default}

\item{coefs}{numeric matrix of simulated log2 means. Must have g rows and k columns, for each gene and cluster. If NULL, must input beta, p_disc, and LFC}

\item{beta}{numeric or numeric vector, gene ``baseline" log2 mean. If not vector, then all genes will be simulated with same baseline}

\item{p_disc}{numeric value from 0 to 1, proportion of genes that are discriminatory (that LFC is applied).}

\item{LFC}{numeric value greater than 0, log2 fold change to be applied on discriminatory genes. For each gene, one random cluster simulated to be either up-regulated (add LFC) or down-regulated (subtract LFC)}

\item{disp}{string, either "gene" or "cluster". "gene" simulates gene-level overdispersion parameters, and "cluster" simulates cluster-level overdispersion parameters}

\item{phi}{numeric, overdispersion parameters. must be vector of length g if disp=='gene', or matrix of g rows and k columns if disp=='cluster'}

\item{size_factors}{numeric vector of simulated size factors for each subject. By default, simulated from N(1,0.25)}

\item{batch_effects}{numeric vector of effects corresponding to each level of 'batch'. Must have length equal to the number of batches specified. Effects are applied in order (i.e. first element of batch_effects refers to the effect of batch 1, etc)}

\item{batch}{integer vector of simulated batch of each sample. Must be numeric 1 to (number of batches).}
}
\value{
list containing the following objects:
y: count matrix of g by n
clusters: cluster indices for each of the n samples
batch: input known batch of each sample (all 1 by default)
batch_effects: input batch effects for each batch (0 by default)
batch_g: genes on which batch effects were applied (on random half of genes)
size_factors: simulated or input size factors (by default, simulated from N(1,0.25))
coefs: matrix of log2 means for each gene and cluster, either input matrix or constructed based on beta, p_disc, and LFC
phi: numeric overdispersion parameters, either vector or matrix based on dispersion scheme
true_disc: logical vector of length g, TRUE if gene is simulated discriminatory (not simulated by same beta across clusters), FALSE if not
}
\description{
Simulate data based on input simulation parameters. Size factors are custom input
or simulated from N(1,0.25)
}
