% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSCseq.R
\name{theta.ml2}
\alias{theta.ml2}
\title{Theta estimation with penalty}
\usage{
theta.ml2(y, mu, n = sum(weights), weights, t0 = 0, limit = 10,
  eps = .Machine$double.eps^0.25, trace = FALSE)
}
\arguments{
\item{y}{cts matrix}

\item{mu}{mean parameters}

\item{n}{number of samples}

\item{weights}{number of weights}

\item{t0}{initial estimate of dispersion scale parameter}

\item{limit}{max number of iterations}

\item{eps}{tolerance}

\item{trace}{boolean whether to trace progress}
}
\value{
theta
}
\description{
theta.ml() with small ridge penalty to stabilize estimates, and using more efficient
digamma and trigamma functions.
}
