% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSCseq.R
\name{EM_run}
\alias{EM_run}
\title{EM/CEM run for FSCseq}
\usage{
EM_run(ncores, X = NA, y, k, lambda = 0, alpha = 0,
  size_factors = rep(1, times = ncol(y)), norm_y = y,
  true_clusters = NA, true_disc = NA, init_parms = FALSE,
  init_coefs = matrix(0, nrow = nrow(y), ncol = k),
  init_phi = matrix(0, nrow = nrow(y), ncol = k), init_cls = NULL,
  init_wts = NULL, CEM = F, init_Tau = 1, maxit_EM = 100,
  maxit_IRLS = 50, EM_tol = 1e-06, IRLS_tol = 1e-04, disp,
  trace = F, optim_method = "direct", mb_size = NULL, BIC_penalty,
  gamma = NULL)
}
\arguments{
\item{X}{design matrix of dimension n by p}

\item{y}{count matrix of dimension g by n}

\item{k}{integer, number of clusters}

\item{lambda}{numeric penalty parameter, lambda >= 0}

\item{alpha}{numeric penalty parameters, 0 <= alpha < 1}

\item{size_factors}{numeric vector of length n, factors to correct for subject-specific variation of sequencing depth}

\item{norm_y}{count matrix of dimension g by n, normalized for differences in sequencing depth}

\item{true_clusters}{(optional) integer vector of true groups, if available, for diagnostic tracking}

\item{true_disc}{(optional) logical vector of true discriminatory genes, if available, for diagnostic tracking}

\item{init_parms}{logical, TRUE: custom parameter initializations, FALSE (default): start from scratch}

\item{init_coefs}{matrix of dimension g by k, only if init_parms = TRUE}

\item{init_phi}{vector of dimension g (gene-specific dispersions) or matrix of dimension g by k (cluster-specific dispersions), only if init_parms = TRUE}

\item{init_cls}{vector of length n, initial clustering. If NA (default), multiple initializations will be searched. init_wts or init_cls must be initialized}

\item{init_wts}{matrix of dim k x n: denotes cluster memberships, but can have partial membership. init_wts or init_cls must be initialized}

\item{CEM}{logical, TRUE for EM (default), FALSE for CEM}

\item{init_Tau}{numeric, initial temperature for CEM. Default is 1 for EM and g for CEM}

\item{maxit_EM}{integer, maximum number of iterations for full EM/CEM run (default 100)}

\item{maxit_IRLS}{integer, maximum number of iterations for IRLS algorithm, in M step (default 50)}

\item{EM_tol}{numeric, tolerance of convergence for EM/CEM, default is 1E-8}

\item{IRLS_tol}{numeric, tolerance of convergence for IRLS, default is 1E-6}

\item{disp}{string, either "gene" (default) or "cluster"}

\item{trace}{logical, TRUE: output diagnostic messages, FALSE (default): don't output}

\item{optim_method}{string, three options "direct", "NR", or "GD". Direct, Newton-Raphson, or Gradient descent (fixed step size of 2)}
}
\value{
FSCseq object: list containing outputs
k: integer order,
pi: numeric vector of mixture proportions,
coefs: numeric matrix of dimensions g by (k+p) cluster log2 baselines and covariate effects,
Q: numeric vector of Q function values across iterations,
BIC: numeric BIC value,
discriminatory: logical vector of length g of whether a gene is discriminatory,
init_clusters: numeric vector of clusters at beginning of EM/CEM algorithm,
init_coefs: matrix of dimension g by k only output if initial parameter estimates were specified,
init_phi: vector of length g or matrix of dimension g by k only output if initial parameter estimates were specified,
final_clusters: vector of length n of resulting clusters,
phi: dispersion estimates,
logL: total log-likelihood,
wts: k by n matrix of E step weights,
time_elap: amount of time elapsed (in seconds),
lambda: input lambda,
alpha: input alpha,
size_factors: input size factors,
norm_y: input norm_y,
DNC: 1 if EM didn't converge in maxits and 0 if it did,
LFCs: matrix of dimension g by maxit_EM of LFCs (max-min)/(k-1)
}
\description{
Performs clustering, feature selection, and estimation of parameters
using a finite mixture model of negative binomials
}
