% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSCseq.R
\name{FSCseq_predict}
\alias{FSCseq_predict}
\title{Prediction via FSCseq}
\usage{
FSCseq_predict(X = NULL, fit, cts_pred, SF_pred)
}
\arguments{
\item{X}{(optional) design matrix of dimension n by p}

\item{fit}{FSCseq results object}

\item{cts_pred}{prediction/test count data matrix of dimension g by n_pred}

\item{SF_pred}{vector of length n_pred, size factors for prediction subjects}
}
\value{
list containing outputs
final_clusters: vector of length n of resulting clusters,
wts: k by n matrix of E step weights
}
\description{
Performs prediction of cluster membership via trained model
}
