% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSCseq_workflow.R
\name{FSCseq_predict_workflow}
\alias{FSCseq_predict_workflow}
\title{Minimal workflow for FSCseq_predict}
\usage{
FSCseq_predict_workflow(
  res,
  X_covar_train = NULL,
  cts_train,
  SF_train = NULL,
  batch_train = NULL,
  X_covar_pred = NULL,
  cts_pred,
  batch_pred = NULL,
  coding = "reference"
)
}
\arguments{
\item{res}{Fitted FSCseq result object.}

\item{X_covar_train}{Optional covariate matrix for training samples (optional additional covariates, except batch)}

\item{cts_train}{Counts matrix of training samples.}

\item{SF_train}{Size factors of training samples (optional. If not supplied, can be accessed from res)}

\item{batch_train}{Batch information for training samples (optional).}

\item{X_covar_pred}{Optional covariate matrix for prediction samples (optional additional covariates, except batch)}

\item{cts_pred}{Counts matrix of prediction samples. Should be same dimension as cts_train}

\item{batch_pred}{Batch information for prediction samples (optional). In estimating the new prediction batch effect, we recommend users to have at least 3-5 samples per cluster per batch for accurate estimation.}

\item{coding}{Coding scheme for batch (categorical). Default is reference coding.}
}
\value{
list with processed.dat.pred (processed prediction data), and prediction results
}
\description{
Full FSCseq workflow based on minimal working defaults
}
\examples{
sim.dat = simulateData(B=1, g=10000, K=2, n=50, LFCg=1, pDEg=0.05, beta0=12, phi0=0.35, nsims=1, save_file=F)[[1]]
\dontrun{res = FSCseq_workflow(cts=sim.dat$cts, K_search=c(2:3), lambda_search=c(1.0, 1.5), alpha_search=c(0.1, 0.2))}
\dontrun{pred_results = FSCseq_predict_workflow(res=res, cts_train=sim.dat$cts_train, batch_train=batch, cts_pred=sim.dat$cts_pred, batch_pred=batch_pred)}

}
\references{
\url{https://github.com/DavidKLim/FSCseq}
}
\author{
David K. Lim, \email{deelim@live.unc.edu}
}
