% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSCseq_workflow.R
\name{FSCseq_predict_workflow}
\alias{FSCseq_predict_workflow}
\title{Minimal workflow for FSCseq_predict}
\usage{
FSCseq_predict_workflow(
  res,
  X_covar_train = NULL,
  cts_train,
  SF_train = NULL,
  batch_train = NULL,
  X_covar_pred = NULL,
  cts_pred,
  batch_pred = NULL,
  coding = "reference"
)
}
\arguments{
\item{cts_pred}{integer matrix, count matrix of dimension g by n_pred. Must be integers (counts)}

\item{X}{covariates (optional)}

\item{fit}{FSCseq results object. Accessed by $results from FSCseq_workflow object}

\item{cts}{Training set count matrix, dimension g by n. Used as pseudo-reference to calculate prediction set size factors}

\item{idx}{boolean vector: TRUE if gene passed pre-filtering step}
}
\value{
list with processed.dat.pred (processed prediction data), and prediction results
}
\description{
Full FSCseq workflow based on minimal working defaults
}
\examples{
sim.dat = simulateData(B=1, g=10000, K=2, n=50, LFCg=1, pDEg=0.05, beta0=12, phi0=0.35, nsims=1, save_file=F)[[1]]
\dontrun{FSCseq_results = FSCseq_workflow(cts=sim.dat$cts, K_search=c(2:3), lambda_search=c(1.0, 1.5), alpha_search=c(0.1, 0.2))}
\dontrun{proc.dat = FSCseq_results$processed.dat}
\dontrun{pred_results = FSCseq_predict_workflow(fit=FSCseq_results$results$fit, cts=sim.dat$cts, cts_pred=sim.dat$cts_pred, idx=proc.dat$idx)}

}
\references{
\url{https://github.com/DavidKLim/FSCseq}
}
\author{
David K. Lim, \email{deelim@live.unc.edu}
}
