% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSCseq_workflow.R
\name{FSCseq_workflow}
\alias{FSCseq_workflow}
\title{Minimal workflow for FSCseq}
\usage{
FSCseq_workflow(cts, ncores = 1, batch = rep(1, ncol(cts)),
  true_cls = NULL, true_disc = NULL, method = "EM", n_rinits = 20,
  med_filt = 500, MAD_filt = 50, K_search = c(2:6),
  lambda_search = seq(0.25, 3, 0.25), alpha_search = c(0.01, seq(0.05,
  0.5, 0.05)), OS_save = TRUE, trace = F, trace.file = NULL,
  nMB = 5)
}
\arguments{
\item{cts}{integer matrix, count matrix of dimension g by n. Must be integers (counts)}

\item{ncores}{integer, number of cores (for parallel computing). Default is 1}

\item{batch}{vector of batch, to use as covariates. Default is no batches.}

\item{true_cls}{(optional) integer vector of true groups, if available, for diagnostic tracking.}

\item{true_disc}{(optional) logical vector of true discriminatory genes, if available, for diagnostic tracking.}

\item{method}{string, either "EM" or "CEM". Default is "EM"}

\item{n_rinits}{integer, number of additional random initializations (on top of Hierarchical and K-means) to be searched. Default is 20}

\item{med_filt}{numeric, threshold for minimum median gene normalized count for pre-filtering. med_filt=0 pre-filters no genes via this criterion. Default is 500.}

\item{MAD_filt}{numeric, value between 0 and 100. quantile threshold for gene log MAD of normalized count. MAD_filt=0 pre-filters no genes via this criterion. Default is 50.}

\item{K_search}{integer vector, values of K (number of clusters) to be searched. Default is 2:6}

\item{lambda_search}{numeric vector, values of lambda to be searched. Default is seq(0.25,3,0.25)}

\item{alpha_search}{numeric vector, values of alpha to be searched. Default is c(0.01,seq(0.05,0.50,0.05))}

\item{OS_save}{logical, TRUE: saves progress of computationally costly warm starts (multiple initializations). Default is TRUE}

\item{trace}{logical, TRUE: output diagnostic messages, FALSE (default): don't output}

\item{trace.file}{(optional) string, file name to store trace output. For example: 'test.txt'}

\item{nMB}{integer, number of minibatches to use in M step. Default is 5}
}
\value{
list with K, cls, discriminatory, and fit
}
\description{
Full FSCseq workflow based on minimal working defaults
}
