% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateData.R
\name{simulateData}
\alias{simulateData}
\title{Simulate RNA-seq bulk gene expression count data}
\usage{
simulateData(K, B = 1, g = 10000, n, pK = NULL, pB = NULL, LFCg,
  pDEg, sigma_g = 0.1, LFCb = 1, pDEb = 0.5, sigma_b = 0, beta0,
  phi0, SF = NULL, nsims = 25, disp = "gene", n_pred = 25,
  save_file = TRUE, save_dir = NULL, save_pref = NULL)
}
\arguments{
\item{K}{integer, number of clusters}

\item{B}{integer, number of batches}

\item{g}{integer, number of genes}

\item{n}{integer, number of samples}

\item{pK}{vector of length K (optional): proportion of samples in each cluster}

\item{pB}{vector of length B (optional): proportion of samples in each batch}

\item{LFCg}{numeric, LFC for cluster-discriminatory genes}

\item{pDEg}{numeric, proportion of genes that are cluster-discriminatory}

\item{sigma_g}{numeric, Gaussian noise added from N(0,sigma_g). Default is 0.1}

\item{LFCb}{numeric, LFC for genes that are differentially expressed across batch. Default is 1.}

\item{pDEb}{numeric, proportion of genes that are differentially expressed across batch. Default is 0.5.}

\item{sigma_b}{numeric, Gaussian noise added to each batch (turned off, set to 0).}

\item{beta0}{numeric, baseline log2 expression for each gene before LFC is applied}

\item{phi0}{numeric, baseline overdispersion for each gene}

\item{SF}{vector of length n (optional), custom size factors from DESeq2. If NULL, simulated from N(1,0.25)}

\item{nsims}{integer, number of datasets to simulate given the input conditions. Default is 25.}

\item{disp}{string, either 'gene' or 'cluster' to simulate gene-level or cluster-level dispersions. Default is gene-level. Input phi must be g x K matrix if disp='cluster'}

\item{n_pred}{integer, number of samples in simulated prediction dataset. Default is 25}

\item{save_file}{boolean: TRUE (save each set of simulations)}

\item{save_dir}{string (optional): directory to save files. Default: 'Simulations/<sigma_g>_<sigma_b>/B<B>'}

\item{save_pref}{string (optional): prefix of file name to save simulated data to. Default: '<K>_<n>_<LFCg>_<pDEg>_<beta0>_<phi0>'}
}
\value{
if save_file=TRUE, then saved file in '<save_dir>/<save_pref>_sim<1:nsims>_data.RData'. Otherwise, list of length 'nsims', with a sim.dat list object for each simulation
}
\description{
Simulate data based on input simulation parameters. Size factors are custom input
or simulated from N(1,0.25)
}
