% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{summary}
\alias{summary}
\title{Summarize FSCseq clustering results}
\usage{
summary(res, true_cls = NULL, true_disc = NULL)
}
\arguments{
\item{res}{Output of FSCseq clustering analysis}
}
\value{
Summary of FSCseq clustering results: #clusters (K), clusters,
ARI (if true_cls input), TPR and FPR (vs. true cluster-discriminatory genes if true_disc input)
}
\description{
Outputs relevant summary of FSCseq clustering results
}
