% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSCseq.R
\name{theta.ml2}
\alias{theta.ml2}
\title{Theta estimation with penalty}
\usage{
theta.ml2(y, mu, n = sum(weights), weights, limit = 10,
  eps = .Machine$double.eps^0.25, trace = FALSE)
}
\arguments{
\item{v}{a vector or matrix of numbers}
}
\value{
theta
}
\description{
theta.ml() with small ridge penalty to stabilize estimates
}
